+++ src/connections/db/migrations/20251120_000005_create_product_variants_table.ts
@@ -9,7 +9,8 @@ export const migration: Migration = {
         product_id INTEGER REFERENCES products(id) ON DELETE CASCADE,
         variant_type VARCHAR(50) NOT NULL,
         variant_value VARCHAR(100) NOT NULL,
-        price_adjustment DECIMAL(10, 2) DEFAULT 0,
+        -- Điều chỉnh giá theo đơn vị VND, dùng INTEGER để đồng bộ với products.price
+        price_adjustment INTEGER DEFAULT 0,
         stock_quantity INTEGER DEFAULT 0,
         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
       )
